audio_processing.py
from pydub import AudioSegment

# Function to process audio
def process_audio(audio_path):
    audio = AudioSegment.from_wav(audio_path)
    # Perform some audio analysis or feature extraction here
    return audio
