requirements.txt
pandas
numpy
streamlit
sentence-transformers
langchain
milvus
pymilvus
openai
PyPDF2
SpeechRecognition
ffmpeg